refactor classes to support subpatterns
