+++ tests/RegRev/Test/Metacharacter/CharType/BlankTest.php
@@ -8,25 +8,24 @@
  * please view the LICENSE file that was distributed with this source code.
  */
 
-namespace RegRev\CharType;
+namespace RegRev\Test\Metacharacter\CharType;
+
+use RegRev\Metacharacter\CharType\Blank;
 
 /**
  * Class Number
  *
  * @package RevReg\Char
  */
-class NonAlnum extends CharType
+class BlankTest extends \PHPUnit_Framework_TestCase
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function generate()
+    public function setup()
     {
-        $characters = "!@£$%^&*()-_=+{}[];:'\"|<>?,./`~";
-        $randomIndex = rand(0, strlen($characters));
+        $this->blank = new Blank();
+    }
 
-        return $characters[$randomIndex];
+    public function testGenerate()
+    {
+        $this->assertEquals(' ', $this->blank->generate());
     }
 }
-
-
