+++ tests/RegRev/Test/Metacharacter/CharType/NonAlnumTest.php
@@ -8,24 +8,24 @@
  * please view the LICENSE file that was distributed with this source code.
  */
 
-namespace RegRev\Test\Char;
+namespace RegRev\Test\Metacharacter\CharType;
 
-use RegRev\CharType\Blank;
+use RegRev\Metacharacter\CharType\NonAlnum;
 
 /**
  * Class Number
  *
  * @package RevReg\Char
  */
-class BlankTest extends \PHPUnit_Framework_TestCase
+class NonAlnumTest extends \PHPUnit_Framework_TestCase
 {
     public function setup()
     {
-        $this->blank = new Blank();
+        $this->nonAlnum = new NonAlnum();
     }
 
     public function testGenerate()
     {
-        $this->assertEquals(' ', $this->blank->generate());
+        $this->assertFalse(ctype_alnum($this->nonAlnum->generate()));
     }
 }
